aula 4

+++ database/migrations/2020_03_11_002457_create_client_table.php
@@ -16,7 +16,7 @@ class CreateClientTable extends Migration
         Schema::create('clients', function (Blueprint $table) {
             $table->unsignedBigInteger('id',true);
             $table->string('name');
-            $table->integer('CPF');
+            $table->BigInteger('CPF');
             $table->string('Email');
             $table->boolean('active_flag')->default(true);
             $table->string('endereco')->nullable();
